dir improvements: move `ethstats` to `node`
